+++ sdk/cosmosdbforpostgresql/azure-resourcemanager-cosmosdbforpostgresql/src/main/java/com/azure/resourcemanager/cosmosdbforpostgresql/fluent/models/ClusterServerInner.java
@@ -6,7 +6,6 @@ package com.azure.resourcemanager.cosmosdbforpostgresql.fluent.models;
 
 import com.azure.core.annotation.Fluent;
 import com.azure.core.management.ProxyResource;
-import com.azure.core.management.SystemData;
 import com.azure.resourcemanager.cosmosdbforpostgresql.models.ServerRole;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -19,12 +18,6 @@ public final class ClusterServerInner extends ProxyResource {
     @JsonProperty(value = "properties")
     private ClusterServerProperties innerProperties;
 
-    /*
-     * Azure Resource Manager metadata containing createdBy and modifiedBy information.
-     */
-    @JsonProperty(value = "systemData", access = JsonProperty.Access.WRITE_ONLY)
-    private SystemData systemData;
-
     /** Creates an instance of ClusterServerInner class. */
     public ClusterServerInner() {
     }
@@ -36,15 +29,6 @@ public final class ClusterServerInner extends ProxyResource {
      */
     private ClusterServerProperties innerProperties() {
         return this.innerProperties;
-    }
-
-    /**
-     * Get the systemData property: Azure Resource Manager metadata containing createdBy and modifiedBy information.
-     *
-     * @return the systemData value.
-     */
-    public SystemData systemData() {
-        return this.systemData;
     }
 
     /**