+++ sdk/cosmosdbforpostgresql/azure-resourcemanager-cosmosdbforpostgresql/src/main/java/com/azure/resourcemanager/cosmosdbforpostgresql/fluent/models/FirewallRuleInner.java
@@ -6,7 +6,6 @@ package com.azure.resourcemanager.cosmosdbforpostgresql.fluent.models;
 
 import com.azure.core.annotation.Fluent;
 import com.azure.core.management.ProxyResource;
-import com.azure.core.management.SystemData;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.cosmosdbforpostgresql.models.ProvisioningState;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -20,12 +19,6 @@ public final class FirewallRuleInner extends ProxyResource {
     @JsonProperty(value = "properties", required = true)
     private FirewallRuleProperties innerProperties = new FirewallRuleProperties();
 
-    /*
-     * Azure Resource Manager metadata containing createdBy and modifiedBy information.
-     */
-    @JsonProperty(value = "systemData", access = JsonProperty.Access.WRITE_ONLY)
-    private SystemData systemData;
-
     /** Creates an instance of FirewallRuleInner class. */
     public FirewallRuleInner() {
     }
@@ -37,15 +30,6 @@ public final class FirewallRuleInner extends ProxyResource {
      */
     private FirewallRuleProperties innerProperties() {
         return this.innerProperties;
-    }
-
-    /**
-     * Get the systemData property: Azure Resource Manager metadata containing createdBy and modifiedBy information.
-     *
-     * @return the systemData value.
-     */
-    public SystemData systemData() {
-        return this.systemData;
     }
 
     /**