+++ sdk/cosmosdbforpostgresql/azure-resourcemanager-cosmosdbforpostgresql/src/main/java/com/azure/resourcemanager/cosmosdbforpostgresql/implementation/ConfigurationImpl.java
@@ -4,7 +4,6 @@
 
 package com.azure.resourcemanager.cosmosdbforpostgresql.implementation;
 
-import com.azure.core.management.SystemData;
 import com.azure.resourcemanager.cosmosdbforpostgresql.fluent.models.ConfigurationInner;
 import com.azure.resourcemanager.cosmosdbforpostgresql.models.Configuration;
 import com.azure.resourcemanager.cosmosdbforpostgresql.models.ConfigurationDataType;
@@ -35,10 +34,6 @@ public final class ConfigurationImpl implements Configuration {
 
     public String type() {
         return this.innerModel().type();
-    }
-
-    public SystemData systemData() {
-        return this.innerModel().systemData();
     }
 
     public String description() {