+++ sdk/cosmosdbforpostgresql/azure-resourcemanager-cosmosdbforpostgresql/src/main/java/com/azure/resourcemanager/cosmosdbforpostgresql/models/FirewallRule.java
@@ -4,7 +4,6 @@
 
 package com.azure.resourcemanager.cosmosdbforpostgresql.models;
 
-import com.azure.core.management.SystemData;
 import com.azure.core.util.Context;
 import com.azure.resourcemanager.cosmosdbforpostgresql.fluent.models.FirewallRuleInner;
 
@@ -30,13 +29,6 @@ public interface FirewallRule {
      * @return the type value.
      */
     String type();
-
-    /**
-     * Gets the systemData property: Azure Resource Manager metadata containing createdBy and modifiedBy information.
-     *
-     * @return the systemData value.
-     */
-    SystemData systemData();
 
     /**
      * Gets the startIpAddress property: The start IP address of the cluster firewall rule. Must be IPv4 format.
@@ -81,11 +73,13 @@ public interface FirewallRule {
             DefinitionStages.WithEndIpAddress,
             DefinitionStages.WithCreate {
     }
+
     /** The FirewallRule definition stages. */
     interface DefinitionStages {
         /** The first stage of the FirewallRule definition. */
         interface Blank extends WithParentResource {
         }
+
         /** The stage of the FirewallRule definition allowing to specify parent resource. */
         interface WithParentResource {
             /**
@@ -97,6 +91,7 @@ public interface FirewallRule {
              */
             WithStartIpAddress withExistingServerGroupsv2(String resourceGroupName, String clusterName);
         }
+
         /** The stage of the FirewallRule definition allowing to specify startIpAddress. */
         interface WithStartIpAddress {
             /**
@@ -108,6 +103,7 @@ public interface FirewallRule {
              */
             WithEndIpAddress withStartIpAddress(String startIpAddress);
         }
+
         /** The stage of the FirewallRule definition allowing to specify endIpAddress. */
         interface WithEndIpAddress {
             /**
@@ -119,6 +115,7 @@ public interface FirewallRule {
              */
             WithCreate withEndIpAddress(String endIpAddress);
         }
+
         /**
          * The stage of the FirewallRule definition which contains all the minimum required properties for the resource
          * to be created, but also allows for any other optional properties to be specified.
@@ -140,6 +137,7 @@ public interface FirewallRule {
             FirewallRule create(Context context);
         }
     }
+
     /**
      * Begins update for the FirewallRule resource.
      *
@@ -164,6 +162,7 @@ public interface FirewallRule {
          */
         FirewallRule apply(Context context);
     }
+
     /** The FirewallRule update stages. */
     interface UpdateStages {
         /** The stage of the FirewallRule update allowing to specify startIpAddress. */
@@ -177,6 +176,7 @@ public interface FirewallRule {
              */
             Update withStartIpAddress(String startIpAddress);
         }
+
         /** The stage of the FirewallRule update allowing to specify endIpAddress. */
         interface WithEndIpAddress {
             /**
@@ -189,6 +189,7 @@ public interface FirewallRule {
             Update withEndIpAddress(String endIpAddress);
         }
     }
+
     /**
      * Refreshes the resource to sync with Azure.
      *