+++ sdk/cosmosdbforpostgresql/azure-resourcemanager-cosmosdbforpostgresql/CHANGELOG.md
@@ -1,14 +1,30 @@
 # Release History
 
-## 1.0.0-beta.2 (Unreleased)
+## 1.0.0-beta.2 (2023-09-15)
 
-### Features Added
+- Azure Resource Manager CosmosDBForPostgreSql client library for Java. This package contains Microsoft Azure SDK for CosmosDBForPostgreSql Management SDK. Azure Cosmos DB for PostgreSQL database service resource provider REST APIs. Package tag package-2022-11-08. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Breaking Changes
 
-### Bugs Fixed
+#### `models.ServerConfiguration` was modified
 
-### Other Changes
+* `systemData()` was removed
+
+#### `models.FirewallRule` was modified
+
+* `systemData()` was removed
+
+#### `models.Configuration` was modified
+
+* `systemData()` was removed
+
+#### `models.ClusterServer` was modified
+
+* `systemData()` was removed
+
+#### `models.Role` was modified
+
+* `systemData()` was removed
 
 ## 1.0.0-beta.1 (2023-06-13)